🔖 Rename : type 폴더 이동 및 쓸모없는 주석 삭제
