--- src/consts/recipe-data.ts
@@ -1,9 +1,0 @@
-import { keyID, serviceID } from "../apis/type/recipe.type";
-
-export const recipeData = {
-  keyId: keyID, // 실제 값으로 대체
-  serviceId: serviceID, // 실제 값으로 대체
-  dataType: "json", // 실제 값으로 대체
-  startIdx: "1", // 시작 인덱스 값
-  endIdx: "12", // 종료 인덱스 값
-};
