+++ src/context/auth.ctx.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useEffect, useState } from 'react';
+import { FC, createContext, useContext, useEffect, useState } from 'react';
 import TokenRepository from '../repositories/token-repository';
 import { AuthContextProps, AuthProviderProps } from '../type/ctx.type';
 
@@ -14,14 +14,8 @@ export const useAuth = (): AuthContextProps => {
     return context;
 };
 
-const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
+const AuthProvider: FC<AuthProviderProps> = ({ children }) => {
     const [accessToken, setAccessToken] = useState(TokenRepository.getToken());
-
-    /* 
-        accessToken을 왜 state로 관리 하는가
-        웹 스토리지는 state가 아니므로, 로그인 로그아웃이 새로 고침이나 페이지 이동 없이
-        UI 변동을 일으키기 위함
-    */
 
     useEffect(() => {
         const token = TokenRepository.getToken();
