--- src/pages/my/dialog.interface.ts
@@ -1,3 +1,0 @@
-export interface MyDialogProps {
-  isDialogOpen: boolean;
-}
