+++ src/type/main.type.ts
